Remove contextual id from sponsored tile telemetry
